grn(1): Drop captions from tables.
No other man page distributed with groff captions tables.

(The way it was done was kind of a kluge, too--I think a better idea
would be to extend the tbl(1) interface to expose captions as distinct
entities rather than simulating them with table continuations.  Then it
would be up to a macro package to place them (and supply hyperlink
anchors), similarly to eqn(1) equations.)