src/roff/troff/input.cpp: Drop misleading comment.
A call of `cleanup_and_exit()` appears to be on every exit path out of
troff, not merely the ones discussed.  If you instrument that function
(or set a breakpoint in a debugger), you can observe it in all of the
following cases.

$ alias tg="./build/test-groff -z"
$ echo | tg
$ echo a | tg
$ printf '.vpt 0\n' | tg
$ printf '.de mac\n.di div\nfoo\n..\n' | tg
$ printf '.de mac\n.di div\nfoo\n..\nbar\n' | tg
$ printf '.de mac\n.di div\nfoo\n.di\n..\nbar\n' | tg
$ printf '.de mac\n.di div\nfoo\n.di\n..\n.mac\n.div\nbar\n' | tg
$ printf '.de mac\n.di div\nfoo\n..\n.wh -1i mac\n' | tg
$ printf '.de mac\n.di div\nfoo\n..\n.wh -1i mac\nbar\n' | tg
troff: error: can't continue page ejection because of current diversion
$ printf '.de mac\n.di div\nfoo\n.di\n..\n.wh -1i mac\nbar\n' | tg

Thus, contra the comment, the code is reached regardless of whether
vertical position traps are enabled, regardless of whether a trap-called
macro starts a diversion and, for that matter, regardless of whether
said trap-called macro also ends the diversion before input ends.