troff(1): Fix typography nits.
When setting a term (a concept or a metasyntactic variable name) in
italics, set suffixed punctuation in italics as well, which is better
(and more idiomatic) dead-tree typography.

However, don't do this for italicized terms like "groff" that are Unix
command names.  I want to leave room for a semantic macro to capture
these in the future.