groff_ms(7): Fix rendering nit.
Stop using 'x' modifier when not otherwise necessary; it draws the rule
crazily, distractingly long on wide terminals.  (This is a judgment
call; if you need the column to expand in the first place, there's no
getting around the fact that some people have insanely wide terminals.
But here, we lack that need.)