groff_man_style(7): Drop claim about \c behavior.
groff 1.22.4's groff(7) page said, "Anything after this escape on the
same line is ignored except \R (which works as usual)."

This wasn't true (see Savannah #61935) and I am entertaining a groff
extension to make it even less true than it is.

Consider the following.

.\" Put args in quotation marks.
.de QUOTE
\[lq]\\$*\[rq]
..
.nf
And indeed there will be time
To wonder, \c
.QUOTE Do I dare?\c
and, \c
.QUOTE Do I dare?
.fi
.pl \n[nl]u

$ nroff -Tascii EXPERIMENTS/prufrock.groff
And indeed there will be time
To wonder, "Do I dare?and, "Do I dare?"