[mdoc]: Stop prefixing center header with "BSD".
Stop prefixing center header with name of operating system.  This was
hard-coded to "BSD" and stuck on all mdoc(7) pages using a default
manual section number, regardless of host operating system.  Bad idea.
Further, the argument to the `Os` call, if any, or a default, is always
disclosed in the left footer, so the foregoing practice gave the reader
no new information.  (On the other hand, header and footer rendering was
not reliable in continuous rendering mode, the default for terminals, in
groff 1.22.4 and earlier.  It is now.)

* tmac/mdoc/doc-common: Do it.

  (Dt): Initialize `doc-volume` as empty and annotate it; calling the
  macro with a numeric second argument overrides the fallback default of
  "LOCAL".  A slash is no longer needed to separate a recognized
  architecture string from the OS name, so drop it.

* tmac/groff_mdoc.7.man: Stop using the string in examples.  Also
  correct table format specification declaring 3 columns but using 2.

* tmac/tests/doc_smoke-test.sh: Update test expectations.