groff_mdoc(7): Update example.
The existing one was hard for me to comprehend; the method attempted for
computing the width of the widest column failed due to problems explored
at length recently on the groff mailing list.  Use tbl(1).  Simulate
output more realistically.

https://lists.gnu.org/archive/html/groff/2022-07/msg00213.html
https://lists.gnu.org/archive/html/groff/2022-08/msg00079.html

For a possibly exhaustive exploration of the underlying issues, see the
following.

https://lore.kernel.org/linux-man/62c1b6748d2faa6263264b9fcaa064495357441b.1658195739.git.nabijaczleweli@nabijaczleweli.xyz/T/#m83835d556f7dd115590ff756305734ce5fb27a18

This commit will provoke warnings when the groff_mdoc(7) page is
rendered to the terminal.

troff:./build/tmac/groff_mdoc.7:930: warning: cannot select font 'CR'

This is because our mdoc implementation needs to remap these fonts as
our man implementation does.

This update does exhibit Savannah #62471.  That's a shame.  It's still
perfectly readable, though.