groff_mdoc(7): Ditch the table of contents.
(1) No other man page we ship presents such a thing; (2) no man page
should be so lengthy or elaborate that one is required; (3) coding one
this way presents obvious DRY problems; and (4) when rendering man(7)
and mdoc(7) to PDF with groff, you get hierarchical navigation links
that correspond to section and subsection headings anyway, a superior
solution to the problem of navigability.

Also ditch the "troff idiosyncrasies" section heading.  It isn't
accurate; what follows is an introduction to roff language syntax, not
its "idiosycrasies".