[docs]: Fix "modulo" solecism.
"Modulus" is a noun and "modulo" an adjective[1], analogous to
"multiplication" and "times", respectively.  (Okay, "modulus" and
"multiplication" are _also_ adjectives, thus "multiplication operator".)
Reportedly, the OED identifies "plus", "minus", "times" and so on as
prepositions.

To summarize, one says "5 {plus,minus,times,divided by,modulo} 2 is..."
but "we now introduce the operation of
{addition,subtraction,multiplication,division,modulus}".

Where's Bertrand Russell when you need him?

[1] arguably, and a Web search will reveal just how intransigent the
    argument is