[docs]: Improve metasyntactic variable discipline.
Try to reduce confusion between strings (variables) and string contents.
Say "contents" when we mean what we would call a string literal in C.
This way, when we speak of "string interpolation", the reader knows we
always mean something that has an identifier.

Refer to to character identifiers as "c" rather than "g"; I find it more
coherent to not regard something as a glyph until it hits intermediate
output, at the earliest, which means we don't deal with with glyphs in
the roff language per se.  (What we do is describe them using
characters, typefaces, sizes, slant, color, and so on, with sufficient
precision that an output device can _produce_ a glyph.)