[devpdf,gropdf]: "use warnings" instead of '-w'.
* font/devpdf/util/BuildFoundries.pl:
* src/devices/gropdf/gropdf.pl:
* src/devices/gropdf/pdfmom.pl: Replace use of '-w' in shebang line with
  "use warnings;".

The shebang's max length is usually 128 as defined in
/usr/include/linux/binfmts.h:
  #define BINPRM_BUF_SIZE 128

There would be errors when @PERL@ is longer than 128, use
'/usr/bin/env perl' can fix the problem, but '/usr/bin/env perl -w'
doesn't work:

/usr/bin/env: perl -w: No such file or directory

So replace "perl -w" with "use warnings" to make it work.

Reported-by: Alexander Kanavin <alex@linutronix.de>

Note from Alex: the Yocto project has been carrying this patch for
a few years, hope it's useful and acceptable. This submission is
a part of 'patch cleanup' where we try to reduce the amount of custom
patches by sending upstream what is appropriate.