groff_mdoc(7): Fix bad rendering and bad advice.
The mdoc package's neurotic aversion to the italic face has an exception
that does not redeem it: it sets metasyntactic ellipses in them, namely
when the `Ar` macro is called without arguments (a magical usage).

"Explicit is better than implicit." -- https://peps.python.org/pep-0020/

Cure the man page of several cases of italic ellipsitis, and tell users
how to avoid contracting this repugnant malady in their own man pages.