doc/groff.texi: Drop some material.
There's a lengthy jeremiad here that (1) characterizes groff as a bit
old-fashioned (not the likeliest selling point) and (2) rails against
the fashionability of WYSIWYG apps.  While this may have been timely in
1999, it's pretty dated now.  According to the Google Books Ngram
Viewer, as of 2019 the occurrence of this term had fallen by about 80%
from its 1990-1993 peak, and returned to a level last seen in 1983.

So just get rid of it.