groff(7), groff(1): Relocate material.
Move table of preprocessor tokens from groff(7), which describes the
language of the GNU troff formatter, to groff(1), which describes the
groff front-end command and offers a synoptic view of the system,
notably including preprocessors.

Without a preprocessor to interpret them, these tokens are simply calls
of (likely undefined) macros.