doc/meref.me.in: Correct \n(po description.
BSD me(7) (presumably Allman) referred to this register's value as the
"simulated page offset", which I didn't find enlightening.  I took out
the "simulated" in an earlier commit, but that is also misleading.  It
is an extra indentation applied only to paragraphs.  Rename it
"paragraph offset".  Fortunately it is accurately described under `.lp`
without reference to this mnemonic.