NEWS: Add item regarding tbl and line numbering.
$ cat ~/tmp/line-numbered-table.roff
.nm 1
Here is a line of output.
Sic transit adispicing meatballs.
Let's pad it out with more content to ensure the line breaks.
.TS
L.
This is my table.
There are many like it but this one is mine.
T{
My table is my best exhibit;
I will guard it as I guard my life.
Even as it requires multiple output lines for an entry.
T}
.TE
What is the line number now?

I defy anyone to defend this:

Unix Version 7 on a SIMH PDP-11/45 (and Heirloom Doctools troff today):

  1 Here is a line of  output.   Sic  transit  adispicing  meatballs.
  2 Let's pad it out with more content to ensure the line breaks.
  6 This is my table.
  7 There are many like it but this one is mine.
  8
  9   3 My table is my best exhibit; I will guard it
 10   4 as  I  guard  my  life.  Even as it requires
 11   5 multiple output lines for an entry.
 12 What is the line number now?