tbl(1): Fix style nits.
Be more consistent about applying a grammatical idiom when presenting
descriptions of language elements.  Start with a sentence in the
imperative mood to capture the sense of user intent or command.  Then
explain details of operation with declarative sentences, often (but not
always) in the passive voice.  I feel this is important to distinguish
operations that are taken on the user's behalf from those we instruct a
user to undertake to achieve the effect they desire.