[tests]: Fix potential problem(s) in trap handler.
* src/roff/groff/tests/smoke-test_html_device.sh: Move call of `cleanup`
  function from end of script to after the tests that generate temporary
  files have been run, and before script possibly exits with status 77
  to skip some checks.  Within trap handler, mask trapped signals while
  running handler.

  (cleanup): Reset trap dispositions to default after removing files;
  once the temporary files are gone, we no longer need a trap handler.
  One might already be running however, and this ensures that its
  "suicide" ("kill -s INT $$") will succeed.

This delicate dance is why I try not to create temporary files in test
scripts.  But grohtml does it for us, so we have to cope.  (Not so
temporary, in fact--you have to keep them around if you want to view the
HTML document produced.)

(To test this stuff, add a "sleep" built-in before the "cleanup" in the
main logic of the script, and kill it.)