[build]: Don't treat Ghostscript fonts as URW.
[build]: Stop scraping output of Ghostscript executable with "-h" option
to attempt to find URW fonts.  Fonts that ship with Ghostscript are
regarded as the "default" foundry, not the URW foundry (though they
often ultimately originate with URW fonts).  They are often missing
Adobe Font Metric (AFM) files, so it is impossible for groff to generate
font description files for them at build time.

* m4/groff.m4 (GROFF_URW_FONTS_CHECK): Drop `AC_REQUIRE` on
  `GROFF_AWK_PATH`.  Drop awk-based scraping of Ghostscript "-h" output.
  Annotate need for sync between this list of characteristic font file
  names and the one in BuildFoundries.

Thanks to Deri James for the ongoing discussions.