[tmac]: Add Spanish language support (3/5).
* tmac/hyphen.es:
* tmac/es.tmac: Update editor aids.

It doesn't seem to be necessary to specify the file encoding for these
files; GNU Emacs and Vim have no trouble figuring it out.  It helps that
(unlike the French localization files), these use no code points that
differ between Latin-1 and Latin-9, so an assumption of either encoding
works out fine.