[troff]: Skip allocation of zero-length arrays.
* src/roff/troff/input.cpp (temp_iterator::temp_iterator): Skip
  allocation of zero-length arrays.  Resolves "-Walloc-zero" warning
  from GCC.

Fixes <https://savannah.gnu.org/bugs/?62398>.  Thanks to Bjarni Ingi
Gislason for the report.

It is not necessary to make conditional the subsequent `delete[]` of a
null pointer.  "If the _delete-expression_ calls the implementation
deallocation function (3.7.3.2), and if the operand of the delete
expression is not the null pointer constant, the deallocation function
will deallocate the storage referenced by the pointer thus rendering the
pointer invalid" (ISO/IEC 14882-1998, §5.3.5, paragraph 4).  Or as
Stroustrup puts it, "Applying _delete_ to zero has no effect." (_The C++
Programming Language, Special Edition_, p. 128).

Also annotate some null pointers with `nullptr` comments to ease any
future transition to C++11, which defines it as a keyword.