[docs]: Reduce use of term "entity".
Doug McIlroy noted this vague term, which groff employs for multiple
purposes.  Eliminate its application to input processing.  There is now
no longer such a thing as an "entity" in the groff language.

* doc/groff.texi (Character Translations): Do it.  Also clarify
  "nothing" as "the dummy character".

  (Using Symbols): Do it.  Also recast explanation of difference between
  characters and glyphs.  Explicitly state that spaces aren't glyphs.
  Document that `rchar` request can't remove definitions supplied by
  font description files.

  (Ligatures and Kerning): Speak of "special characters", not
  "entities".

  (Other Differences): Recast discussion of character-to-glyph
  transformation.  Stop qualifying characters as "input".  Recast
  discussion of example.

* font/devutf8/NOTES: Revise use of terminology.  Perform a Kemper
  notectomy.  Wrap long lines.

* man/groff.7.man (Request short reference) <char>: Speak of a "special
  character", not an "entity".

  <rchar>: Document that request can't remove definitions supplied by
  font description files.

* man/groff_diff.7.man (Implementation differences): Sync with our
  Texinfo manual.

The use of "entity" to describe how a glyph gets mapped back to a
character (sequence) for the HTML and terminal output devices is
retained.  That usage is restricted to discussion of output drivers
(code comments and function names notwithstanding).