groff_mm(7): De-document `Indcmd` string.
It is defined in the macro package source but never interpolated.  It
appears to have been replaced between groff 1.14 (December 1999) and
"pre-1.16" (February 2000) with mmroff(1).

Also, it was incorrectly described even 24 years ago; the man page
documented arguments (to sort(1)) as default, but the string definition
in the package did not.