NEWS: Clarify claim about mandoc.
mandoc _does_ permit the line length to be configured (for its "ascii"
output mode only) via a command-line option ("-O width=nn").  This does
not constitute support for doing so via the `ll` request or `LL`
register, so the availability of this feature does not constitute a
barrier to, or incompatibility with, groff's removal of the hacky
feature in question.

(Heirloom Doctools man/troff _does_ honor `ll` requests, but the macro
package changes the line length setting back to its default at various
times, which I don't regard as "support" for changing the formatted line
length.)