eqn(1): Remove some bad advice.
...dating back to commit 62248e78a3, 2007-02-06.  "cdot cdot cdot" is
not equivalent to "cdots" because the two input sequences have different
spacing types.  `cdot` is spaced like a binary operator (which it often
is, as in the dot product), and `cdots`, being an ellipsis on the math
axis, uses the "inner" spacing type.

Also refer to an ellipsis on the text baseline in those terms.