doc/groff.texi: Revise first chapter.
Most of this was wordsmithery and style tweaking, but I reduced and
updated the GNU troff "capability" list to (1) make it less
underwhelming to the newcomer and (2) focus on perceptible aspects of
the typesetting.  As it was, it was too much of a recapitulation of the
contents of the GNU troff reference chapter.  All the jargon about the
formatter's programming language features (registers, diversions, traps,
environments, and so on) is practically meaningless at this point in the
presentation, and in any case they only exist _to facilitate
typesetting_.  So I recast with that objective in mind.

Incidentally, this revision brings the U.S. letter page count down by 2.