doc/groff.texi: Bump version and date.
Documentation changes can be cherry-picked to a stable branch (or
distributor release).  But this document as a whole does not accurately
describe groff 1.23.0, so stop claiming that it does.

We could solve this and the command prefix problem ("gtroff", etc.) by
generating "groff.texi" from "groff.texi.in", substituting in the Git
version information and the command prefix used in the actual build.
Feedback welcome.