groff_man_style(7): Tighten wording.
This un-strands the last line of the "Authors" paragraph.

This problem underscores the essential limitation of the "poor man's
keep" (.br, .ne).  When the end of a paragraph rather than the beginning
is liable to stranding, the `ne` approach forces the whole thing to be
set on a single page instead of breaking it late (but before its last
output line).

Urge to fix Savannah #62911...RISING!

Also improve grammatical parallelism.