[gxditview]: Add accelerators for "Print" action.
An apparent bug in a support library affects some users; it disables
gxditview's pop-up menu.  (The menu pops up, but no item can be
selected.)  Also, "Print" was the only action for which no keyboard
accelerator was defined.  "p" (unmodified) is already bound to
`PreviousPage()`; bind "p" with any of the modifiers Shift, Ctrl, Alt,
or Meta to `Print()`.

* src/devices/xditview/GXditview.ad: Do it.
* src/devices/xditview/gxditview.1.man: Document it.