[docs]: Improve discussion of `fzoom` request.
What does it really mean for this request to be "transparent" to the
formatter?  ("Transparent" has to be one of the most overloaded, abused,
and unhelpful words in groff's documentary corpus.  I am starting to
think it is a placeholder term for anything the writer couldn't figure
out how to explain.)

Also, put the example of usage after the full discussion of request
behavior, including input validity constraints.  Improve the example to
illustrate a _practical_ application of this request.

Try to clarify.  A telling experiment follows.

$ cat EXPERIMENTS/fzoom-width.groff
The width of foo is \w'foo' at \n[.ps] scaled points.
.sp
.ps 20
.vs 24
The width of foo is \w'foo' at \n[.ps] scaled points.
.sp
.ps 10
.vs 12
.fzoom TR 2000
The width of foo is \w'foo' at \n[.ps] scaled points.
$ groff -a EXPERIMENTS/fzoom-width.groff
<beginning of page>
The width of foo is 13330 at 10000 scaled points.
The width of foo is 26660 at 20000 scaled points.
The width of foo is 26660 at 10000 scaled points.

Bump document date.