[build]: Add/update .gitignore files.
Use a surgical, not a shotgun, approach.  Remove stale and/or overbroad
patterns.  This way it is easier to detect bitrot and junk in the build
tree, and recognize when changes occur that affect payload or might need
to be reflected in Makefile updates.

Organize the files by "module"--by logically discrete components of the
tree.  This makes the internal symmetry of the repository less obscure.

These files' contents reflect an "everything" build on GNU/Linux.  Gaps
are expected for platforms with different file naming
conventions--Cygwin, for instance, still names executables with ".exe"
extensions, and may still use ".obj" instead of ".o".  Please contribute
similar patterns for your platform.