[mm]: Fix Savannah #64336 (lost floats at end).
* contrib/mm/m.tmac: Flush displays at end of input even if they cause
  page breaks.  Inspection of source code suggests that the problem
  dates back to the introduction of the package in groff 1.04,
  1991-11-17.  Our mm stumbled into a pitfall of AT&T compatibility.
  Our Texinfo manual (which did not exist yet in 1991) says: "For
  historical reasons (compatibility with AT&T 'troff'), the end-of-input
  macro exits as soon as it causes a page break if no partially
  collected line remains. ..."

Fixes <https://savannah.gnu.org/bugs/?64336>.

Maybe a better design would have been to require end-of-input traps
(those planted with the `em` request) to explicitly exit with `ex` (or
`ab`) when they finish their business.  And maybe we should make that
change in a future groff release (but keep the old behavior in
"compatibility mode").