doc/webpage.ms: Fix font name solecism.
Request font 'CR', not font 'C', which is the name of a _family_ in
groff, not a resolved font name.

A better fix is likely to write a macro for these ersatz displays, using
ms's displays macros and switching _families_ in the bargain.  However,
there is Savannah #61915 to worry about.