groff_www(7): Revise.
Content:
* Drop stale advice; this page no longer uses these macros itself (and
  has not since groff 1.22.4 or earlier) because that reduces its
  portability as a man page.  "doc/webpage.ms" is a better exemplar.
* Characterize grohtml(1) as being in "beta", not "alpha" state.  (In my
  opinion, if we can't defend a claim of beta quality after 20+ years,
  we should withdraw it.)
* Fix numerous errors in macro syoposes, such as arguments being out of
  order in `IMG`.
* Drop displayed examples.  They illustrated only the obvious.
* Drop documentation of `FTP` macro, which seems entirely pointless.
  (It is aliased to `URL` in the implementation.)

Style:
* Respell "-Txxx" formatter options as "-T xxx".
* Spell macro names without leading dot in summary table and synopses.
  A user of this macro package is expected to know basic call syntax.
* Properly use roman and italic styles in synopses instead of bold for
  everything.
* Tune indentation amount used for tagged paragraph sequence.
* Recast for clarity.
* Tighten.

Markup:
* Use two empty requests where vertical space is expected.
* Annotate some problems and unclear points.

In reviewing this document, I find myself questioning the design of this
package and a fair amount of its chosen mission.  I think we would do
better to integrate specific features (like hyperlinks) into each
package rather than heaping a hodgepodge of HTML features on top of,
theoretically, any macro package that manages page location traps.  This
auxiliary package is over 20 years old and that mission does not appear
to have been taken up.  It seems to have been written with ms(7) in mind
with little thought given to any other "full-service" package.  (Some
effort was made at promoting it as an extension to man(7), but that idea
seems to have been abandoned about 15 years ago, for which we can be
grateful.)

The inconsistency between the argument lists of `IMG` and `PIMG` is
shockingly bad, and a good example of excessive conservatism (or
inertia) in interface design.

The macros are documented in an ad hoc order with no explicit grouping.
The list-related macros (ULS...LI) lack synopses altogether, a strange
defect in a document that not only explained but provided code displays
for blindingly obvious example applications.

The pseudo-Unix-command-line syntax of *roff macro calls with "option"
arguments beginning with `-` is a terrible practice.  *roff is not the
Unix shell.  There are no file descriptors, no redirection, and no
piping.  Research suggests that blame for this analogical aping can be
laid at mm(7)'s doorstep.

A macro interface with a sequence of some optional arguments followed by
some mandatory ones followed by more optional ones is repugnant because
it is difficult to learn.  It also makes argument processing within
macro definitions more tedious and therefore more prone to error,
because the programmer gets fatigued and overlooks meat and potatoes
issues like syntactical and semantic validation.  Lack of validation
then frustrates the user, because they get diagnostic messages they
don't understand, or, worse, incorrect output with no diagnosis of error
at all.  All mandatory arguments should come first.  That anyone thinks
they shouldn't, or that it helps anyone to have them do so, is another
Unix command-line hangover.  And these days, a geriatric one at that.
Decades ago we learned that such a practice makes it harder to compose
Unix command lines (by stuffing arguments into shell variables, for
example), which is one reason the '--' "end of options" convention
arose, as with with GNU getopt()'s automatic resequencing of argv[].