[docs]: Drop mention of terminal tab stop width.
Revert most of 97c911c0875, 2 September.  It doesn't matter to the
formatter where a terminal output device's tab stops are.  That's a
matter for grotty(1) to worry about, and by default (that is, if not
given the `-h` option) even it doesn't care; it uses spaces to align the
output.

Adding terminfo(3) support to grotty (currently in development) promises
opportunities for micro-optimization of output.  And bugs, if terminal
descriptions advertise `hts` and `tbc` capabilities but the device (or
emulator) doesn't correctly implement them--or someone deliberately
chooses a $TERM that fits their terminal (emulator) poorly.  Grasping
randomly at terminal names used to be quite the sport among greenhorn
Unix users; we might find out if it still is.  (More conservatively, we
can keep grotty's use of hard tabs exactly where it is, and require
users to explicitly enable them.)