[troff]: Recast awful "transparently" diagnostic.
* src/roff/troff/input.cpp (token::next, device_macro_request): Recast
  language of one of groff's more notorious diagnostics, eliminating the
  much-abused word "transparently".  Also name the offending token.  It
  is difficult to overstate how unhelpful diagnostic messages are when
  they know precisely what is wrong but refuse to tell the user.

Tested as follows.

$ printf '.devicem br\n\\Y(br\n' | ./build/test-groff -z
troff:<standard input>:1: error: cannot interpolate 'br' to device-independent output; it is a request, not a macro
troff:<standard input>:2: error: cannot interpolate 'br' to device-independent output; it is a request, not a macro

Lengthy beats incomprehensible.

(One might quibble over my choice of preposition; I was torn between
"to" and "in".)